+++ builds/windows/test.r
@@ -1,4 +1,7 @@
-rebol [
-	title: "test script"
-	purpose: {put your test here, and launch "t" to test it}
-	]
+view [ button "kk" 
+banner "test"
+h5  "test 2"
+btn-cancel
+btn-enter
+btn-help
+]